+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2020-07-20
 * objects: fixed a list-watcher direct-editing offset bug
+* morphic: update the Hand's position on mouse-down - avoid triggering at the origin point
 
 ### 2020-07-19
 * blocks: blocks-fade-out support for label arrows (under construction)